TradingPipeline

Documentation for TradingPipeline.

TradingPipeline.simulate_sanity_check_failure_error

TradingPipeline.AbstractExchangeResponse

TradingPipeline.AbstractManualCommand

TechnicalIndicatorCharts.visualize

TradingPipeline.around

TradingPipeline.decide

TradingPipeline.load_strategy

TradingPipeline.load_strategy

TradingPipeline.percent_diff

TradingPipeline.pnl

TradingPipeline.pnl

TradingPipeline.qty

TradingPipeline.qty

TradingPipeline.report

TradingPipeline.simulate

TradingPipeline.@hsm

TradingPipeline.@pnl

TradingPipeline.@pnls

TradingPipeline.simulate_sanity_check_failure_error — Constant

simulate_sanity_check_failure_error is a tuple filled with a lot of nothing values so that code that's @unpack'ing return values from simulate() don't crash.

TradingPipeline.AbstractExchangeResponse — Type

An AbstractExchangeResponse represents a message that was received from an ExchangeFillSubject. The most common is ExchangeFill which lets the StrategySubject know that after putting out an intent to open or close a position, the exchange has filled the order and the requested position change has occurred.

Example

julia> subtypes(TradingPipeline.AbstractExchangeResponse)

TradingPipeline.AbstractManualCommand — Type

An AbstractManualCommand represents a manual intervention from a human to the StrategySubject.

Example

julia> subtypes(TradingPipeline.AbstractExchangeResponse)

TechnicalIndicatorCharts.visualize — Method

visualize(
    t::Tuple{TechnicalIndicatorCharts.Chart, ExchangeOperations.AbstractSession};
    kwargs...
) -> LightweightCharts.LWCLayout

Let's see if I can visualize trades on top of a chart.

TradingPipeline.around — Method

around(
    ts::Dates.DateTime,
    df::DataFrames.AbstractDataFrame;
    before,
    after,
    ts_field
) -> Any

Return a few rows before and after the given timestamp ts in the DataFrame df.

TradingPipeline.decide — Method

decide(strategy::AbstractStrategy, state::HSM.AbstractHsmState)
-> Union{Nothing, TradeDecision.T}

The decide method takes the current strategy and the current state according to the state machine, and looks at the market data available to it to make a decision. Often, it decides to return nothing, but if conditions as defined by the strategy are met, it could return an HSM.AbstractHsmEvent which will cause the state machine to move to the next appropriate state.

There are decide methods for every state in the state machine.

TradingPipeline.load_strategy — Method

Initialize a long-only simple golden cross strategy.

TradingPipeline.load_strategy — Method

Initialize a long-only hma strategy.

Looking for 330/440 crosses

TradingPipeline.percent_diff — Method

percent_diff(a, b) -> Any

Return the percent difference between a and b.

TradingPipeline.pnl — Method

pnl(long::Long) -> TradingPipeline.Result

Calculate profit/loss of a long position.

TradingPipeline.pnl — Method

pnl(short::Short) -> TradingPipeline.Result

Calculate profit/loss of a short position.

TradingPipeline.qty — Method

qty(c::Contracts, price::Number)

Convert contracts into the equivalent quantity of the asset being traded.

TradingPipeline.qty — Method

qty(n::Number, price::Number)

This just returns n since the quantity n.

TradingPipeline.report — Method

report(session::XO.SimulatorSession) -> DataFrame

Return a list of trades that happened during the simulator session.

TradingPipeline.simulate — Method

simulate(
    candle_observable,
    strategy_type::Type{<:TradingPipeline.AbstractStrategy};
    kwargs...
) -> Any

Run a strategy on the simulator using the given candle_observable.

Return Values

A named tuple with the following keys will be returned: simulator_session, hsm, simultator_exchange_driver_subject, fill_observable, chart_subject, strategy_subject, simulator_session_actor.

Example

julia> candle_observable = df_candles_observable(btcusd1m)
IterableObservable(Candle, Vector{Candle}, Rocket.AsapScheduler)

julia> @unpack simulator_session, chart_subject = simulate(candle_observable, HMAStrategy);

TradingPipeline.@hsm — Macro

@hsm strategy_subject

Given a StrategySubject instance, define the states and transitions for a MarketOrderStrategyStateMachine, and return the singleton instance of that state machine.

TradingPipeline.@pnl — Macro

@pnl entry exit quantity leverage

Calculate the profit/loss of a long positon. The quantity and leverage are optional.

TradingPipeline.@pnls — Macro

@pnls entry exit quantity leverage

Calculate the profit/loss of a short positon. The quantity and leverage are optional.